Safari-611.1.10/LayoutTests/js/dom/JSON-stringify-string-object-with-tojson-expected.txt

PASS nativeJSON.stringify(stringObject) is "\"Foo\""
PASS nativeJSON.stringify(stringObject) is JSON.stringify(stringObject)
PASS nativeJSON.stringify(stringObject) is "\"Weird Case 1\""
PASS nativeJSON.stringify(stringObject) is JSON.stringify(stringObject)
PASS nativeJSON.stringify(stringObject) is "\"Bar\""
PASS nativeJSON.stringify(stringObject) is JSON.stringify(stringObject)
PASS nativeJSON.stringify(stringObject) is "\"Weird Case 2\""
PASS nativeJSON.stringify(stringObject) is JSON.stringify(stringObject)
PASS successfullyParsed is true

TEST COMPLETE
